# Backend/bus-backend-api/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebApplication1, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/BusMaster": {
      "get": {
        "tags": [
          "BusMaster"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BusMasterModel"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BusMaster"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BusMasterModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BusMasterModel"
            }
          }
        }
      }
    },
    "/api/BusMaster/{id}": {
      "get": {
        "tags": [
          "BusMaster"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BusMasterModel"
            }
          }
        }
      },
      "put": {
        "tags": [
          "BusMaster"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number",
            "format": "double"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BusMasterModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "BusMaster"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BusRouteStoppage": {
      "get": {
        "tags": [
          "BusRouteStoppage"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BusRouteStoppageDetailsModel"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BusRouteStoppage"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BusRouteStoppageDetailsModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BusRouteStoppage/{id}": {
      "get": {
        "tags": [
          "BusRouteStoppage"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BusRouteStoppageDetailsModel"
            }
          }
        }
      },
      "put": {
        "tags": [
          "BusRouteStoppage"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number",
            "format": "double"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BusRouteStoppageDetailsModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "BusRouteStoppage"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Login/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/StationMaster": {
      "get": {
        "tags": [
          "StationMaster"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/StationMaster"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "StationMaster"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StationMaster"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StationMaster"
            }
          }
        }
      },
      "put": {
        "tags": [
          "StationMaster"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StationMaster"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/StationMaster/{id}": {
      "get": {
        "tags": [
          "StationMaster"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StationMaster"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StationMaster"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/StationRoutes": {
      "get": {
        "tags": [
          "StationRoutes"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "StationRoutes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StationRouteModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/StationRoutes/{id}": {
      "get": {
        "tags": [
          "StationRoutes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "StationRoutes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StationRouteModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "StationRoutes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Stations/get-stations": {
      "get": {
        "tags": [
          "Stations"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "fromStation",
            "type": "string"
          },
          {
            "in": "query",
            "name": "toStation",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User": {
      "get": {
        "tags": [
          "User"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "BusMasterModel": {
      "required": [
        "busNumber"
      ],
      "type": "object",
      "properties": {
        "busid": {
          "format": "double",
          "type": "number"
        },
        "busNumber": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "busType": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "capacity": {
          "format": "int32",
          "type": "integer"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        }
      }
    },
    "BusRouteStoppageDetailsModel": {
      "type": "object",
      "properties": {
        "stopID": {
          "format": "double",
          "type": "number"
        },
        "busID": {
          "format": "double",
          "type": "number"
        },
        "stationID": {
          "format": "double",
          "type": "number"
        },
        "arrivalTime": {
          "type": "string"
        },
        "departureTime": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "doneAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "StationMaster": {
      "required": [
        "district",
        "stationName",
        "tahsil"
      ],
      "type": "object",
      "properties": {
        "stationID": {
          "format": "double",
          "type": "number"
        },
        "stationName": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "revenueName": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "doneAt": {
          "format": "date-time",
          "type": "string"
        },
        "district": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "tahsil": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        }
      }
    },
    "StationRouteModel": {
      "required": [
        "busId",
        "date",
        "departureTime",
        "fromStation",
        "toStation",
        "totalDistance"
      ],
      "type": "object",
      "properties": {
        "busRouteId": {
          "format": "double",
          "type": "number"
        },
        "busId": {
          "format": "double",
          "type": "number"
        },
        "fromStation": {
          "maxLength": 255,
          "minLength": 1,
          "type": "string"
        },
        "toStation": {
          "maxLength": 255,
          "minLength": 1,
          "type": "string"
        },
        "viaRouteName": {
          "maxLength": 255,
          "type": "string"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "departureTime": {
          "format": "date-time",
          "type": "string"
        },
        "dropTime": {
          "format": "date-time",
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "modifiedDate": {
          "format": "date-time",
          "type": "string"
        },
        "totalDistance": {
          "format": "int32",
          "type": "integer"
        },
        "daysAvailable": {
          "maxLength": 50,
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "userID": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "modifiedDate": {
          "format": "date-time",
          "type": "string"
        },
        "modifiedBy": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "roleID": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  }
}